requirements.txt
streamlit
streamlit-permalink
numpy
pillow
pandas
importlib_resources
matplotlib
